requirements.txt
torch
torchvision
numpy
matplotlib
scipy
scikit-learn
Pillow
tqdm
jupyter
